+++ plugins/escape.ts
@@ -3,6 +3,8 @@ import type { Environment } from "../src/environment.ts";
 
 export default function () {
   return (env: Environment) => {
-    env.filters.escape = html.escape;
+    // deno-lint-ignore no-explicit-any
+    env.filters.escape = (value: any) =>
+      value ? html.escape(value.toString()) : "";
   };
 }