+++ test/tokenizer.test.ts
@@ -1,4 +1,5 @@
-import tokenize, { parseTag } from "../src/tokenizer.ts";
+import { parseTag } from "../src/tokenizer.ts";
+import tmpl from "../mod.ts";
 import { assertEquals } from "https://deno.land/std@0.205.0/assert/assert_equals.ts";
 
 Deno.test("Parse tag", () => {
@@ -12,7 +13,7 @@ Deno.test("Parse tag", () => {
 
 Deno.test("Basic tokenizer", () => {
   const code = `<h1>{{ message }}</h1>`;
-  const { tokens } = tokenize(code);
+  const tokens = tmpl().tokenize(code);
   assertEquals(tokens, [
     ["string", "<h1>", 0],
     ["tag", "message", 4],
@@ -22,7 +23,7 @@ Deno.test("Basic tokenizer", () => {
 
 Deno.test("Tokenizer (doble quotes)", () => {
   const code = `<h1>{{ message + "{{}}" }}</h1>`;
-  const { tokens } = tokenize(code);
+  const tokens = tmpl().tokenize(code);
   assertEquals(tokens, [
     ["string", "<h1>", 0],
     ["tag", 'message + "{{}}"', 4],
@@ -32,7 +33,7 @@ Deno.test("Tokenizer (doble quotes)", () => {
 
 Deno.test("Tokenizer (single quotes)", () => {
   const code = `<h1>{{ message + '{{"}}' }}</h1>`;
-  const { tokens } = tokenize(code);
+  const tokens = tmpl().tokenize(code);
   assertEquals(tokens, [
     ["string", "<h1>", 0],
     ["tag", "message + '{{\"}}'", 4],
@@ -42,7 +43,7 @@ Deno.test("Tokenizer (single quotes)", () => {
 
 Deno.test("Tokenizer (inner curly brackets)", () => {
   const code = `<h1>{{ message + JSON.stringify({fo: {}}) }}</h1>`;
-  const { tokens } = tokenize(code);
+  const tokens = tmpl().tokenize(code);
   assertEquals(tokens, [
     ["string", "<h1>", 0],
     ["tag", "message + JSON.stringify({fo: {}})", 4],
@@ -52,7 +53,7 @@ Deno.test("Tokenizer (inner curly brackets)", () => {
 
 Deno.test("Tokenizer (inner comment)", () => {
   const code = `<h1>{{ message /* }} */ }}</h1>`;
-  const { tokens } = tokenize(code);
+  const tokens = tmpl().tokenize(code);
   assertEquals(tokens, [
     ["string", "<h1>", 0],
     ["tag", "message /* }} */", 4],
@@ -62,7 +63,7 @@ Deno.test("Tokenizer (inner comment)", () => {
 
 Deno.test("Tokenizer (left trim)", () => {
   const code = `<h1> {{- message }} </h1>`;
-  const { tokens } = tokenize(code);
+  const tokens = tmpl().tokenize(code);
   assertEquals(tokens, [
     ["string", "<h1>", 0],
     ["tag", "message", 5],
@@ -72,7 +73,7 @@ Deno.test("Tokenizer (left trim)", () => {
 
 Deno.test("Tokenizer (right trim)", () => {
   const code = `<h1> {{message -}} </h1>`;
-  const { tokens } = tokenize(code);
+  const tokens = tmpl().tokenize(code);
   assertEquals(tokens, [
     ["string", "<h1> ", 0],
     ["tag", "message", 5],
@@ -82,7 +83,7 @@ Deno.test("Tokenizer (right trim)", () => {
 
 Deno.test("Tokenizer (both trims)", () => {
   const code = `<h1> {{-message -}} </h1>`;
-  const { tokens } = tokenize(code);
+  const tokens = tmpl().tokenize(code);
   assertEquals(tokens, [
     ["string", "<h1>", 0],
     ["tag", "message", 5],
@@ -152,7 +153,7 @@ Deno.test("Tokenizer (preserve both)", () => {
 
 Deno.test("Tokenizer (comment)", () => {
   const code = `<h1> {{# {{ message }} #}} </h1>`;
-  const { tokens } = tokenize(code);
+  const tokens = tmpl().tokenize(code);
   assertEquals(tokens, [
     ["string", "<h1> ", 0],
     ["comment", " {{ message }} ", 5],
@@ -162,7 +163,7 @@ Deno.test("Tokenizer (comment)", () => {
 
 Deno.test("Tokenizer (literal)", () => {
   const code = "<h1>{{ `message {}}` }}</h1>";
-  const { tokens } = tokenize(code);
+  const tokens = tmpl().tokenize(code);
   assertEquals(tokens, [
     ["string", "<h1>", 0],
     ["tag", "`message {}}`", 4],
@@ -172,7 +173,7 @@ Deno.test("Tokenizer (literal)", () => {
 
 Deno.test("Tokenizer (literal 2)", () => {
   const code = "<h1>{{ `message ${ JSON.stringify({o:{}}) }` }}</h1>";
-  const { tokens } = tokenize(code);
+  const tokens = tmpl().tokenize(code);
   assertEquals(tokens, [
     ["string", "<h1>", 0],
     ["tag", "`message ${ JSON.stringify({o:{}}) }`", 4],
@@ -182,7 +183,7 @@ Deno.test("Tokenizer (literal 2)", () => {
 
 Deno.test("Tokenizer (filter)", () => {
   const code = "{{ url |> await fetch |> await json |> stringify }}";
-  const { tokens } = tokenize(code);
+  const tokens = tmpl().tokenize(code);
   assertEquals(tokens, [
     ["string", "", 0],
     ["tag", "url", 0],