+++ test/unescape.test.ts
@@ -15,7 +15,7 @@ Deno.test("Unescape filter", async () => {
   });
 });
 
-Deno.test("Escape by default", async () => {
+Deno.test("Unescape with escape by default", async () => {
   await test({
     options: {
       autoescape: true,
@@ -33,5 +33,35 @@ Deno.test("Escape by default", async () => {
     {{ "<h1>Hello world</h1>" |> unescape |> safe }}
     `,
     expected: "<h1>Hello world</h1>",
+  });
+});
+
+Deno.test("Unescape non-string", async () => {
+  await test({
+    template: `
+    {{ 100 |> unescape }}
+    `,
+    expected: "100",
+  });
+  testSync({
+    template: `
+    {{ 100 |> unescape }}
+    `,
+    expected: "100",
+  });
+});
+
+Deno.test("Unescape undefined", async () => {
+  await test({
+    template: `
+    {{ undefined |> unescape }}
+    `,
+    expected: "",
+  });
+  testSync({
+    template: `
+    {{ undefined |> unescape }}
+    `,
+    expected: "",
   });
 });