+++ CHANGELOG.md
@@ -4,9 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.10.3] - Unreleased
+### Added
+- Preprocessor hooks [#33].
+- Support for query strings and hashes in the import paths [#33].
+
+### Changed
+- Internal: refactor tag trimming.
+
+### Fixed
+- `escape` and `unescape` filters for non-string values [#36].
+
 ## [0.10.2] - 2024-01-16
 ### Fixed
-- `for` not working as expected with AsyncIterable? [#28]
+- `for` not working as expected with AsyncIterable [#28]
 
 ## [0.10.1] - 2024-01-08
 ### Added
@@ -161,7 +172,10 @@ First version
 [#20]: https://github.com/oscarotero/vento/issues/20
 [#25]: https://github.com/oscarotero/vento/issues/25
 [#28]: https://github.com/oscarotero/vento/issues/28
+[#33]: https://github.com/oscarotero/vento/issues/33
+[#36]: https://github.com/oscarotero/vento/issues/36
 
+[0.10.3]: https://github.com/oscarotero/vento/compare/v0.10.2...HEAD
 [0.10.2]: https://github.com/oscarotero/vento/compare/v0.10.1...v0.10.2
 [0.10.1]: https://github.com/oscarotero/vento/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/oscarotero/vento/compare/v0.9.2...v0.10.0